Widen sound-effect handles to uintptr for x64 (x64 fix)
A sound-effect "handle" is really a CSample* stored as a uint32:
bPlaySoundEffect writes `pser->u4SoundID = (uint32)(uintptr)psam`, the
caller keeps it in u4AudioID, then StopSoundEffect((CSample*)u4_id)
reconstructs it. On x64 the CSample* was truncated to 32 bits -> stopping
a dino/Anne vocalization dereferenced a garbage pointer. Widen the whole
handle chain to uintptr: SSoundEffectResult::u4SoundID (SoundDefs.hpp),
the (uintptr)psam store, the free StopSoundEffect param (MsgAudio.hpp
decl + AudioDaemon.cpp def), and the u4AudioID fields (CBrain, CPlayer).
The handles are never serialized ("Don't save me!"), so no save-format
change. uintptr==uint32 on Win32 (build green); x64 clean.

This clears the last dangerous C4312: every truncated-pointer->pointer
reconstruction is now fixed. The only C4312 left is GroffIO 1341/1346,
the intended small material-index-in-pSurface-slot (high bits 0, safe).

+++ jp2_pc/Source/Game/AI/Brain.hpp
@@ -255,7 +255,7 @@ public:
 	TSec					sSampleStart;	//  When we started the sound.
 	TSec					sOpenMouth;		//  When we should open the mouth
 	TSec					sCloseMouth;	//  When we should close the mouth
-	uint32					u4AudioID;	// The ID of the vocalization currently being made by the dino.
+	uintptr					u4AudioID;	// The ID of the vocalization currently being made by the dino.
 										//  Zero if none.
 
 	//
